% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_hex_grids.R
\docType{data}
\name{cr_hex_grid_res_5}
\alias{cr_hex_grid_res_5}
\alias{cr_hex_grid_res_6}
\alias{cr_hex_grid_res_7}
\alias{cr_hex_grid_res_8}
\alias{cr_hex_grid_res_9}
\title{Hierarchical hexagonal geospatial grids for Costa Rica of resolution 5, 6, 7, 8 and 9 (H3, Hexagonal hierarchical geospatial indexing system)}
\format{
An sf object (EPSG:4326)
}
\usage{
cr_hex_grid_res_5
cr_hex_grid_res_6
cr_hex_grid_res_7
cr_hex_grid_res_8
cr_hex_grid_res_9
}
\description{
Hierarchical hexagonal geospatial grids for Costa Rica of resolution 5, 6, 7, 8, and 9 (H3, Hexagonal hierarchical geospatial indexing system).
}
\keyword{datasets}
